Refactor code to use useShouldShowInterrupt hook for showing interrupt button in RunButton and Controls components
